+++ R/lsp_add_spatial.R
@@ -203,7 +203,7 @@ lsp_add_terra = function(x = NULL, window = NULL){
   if (!requireNamespace("terra", quietly = TRUE)){
     stop("package terra required, please install it first") # nocov
   }
-  output = lsp_add_stars(x = x, window = window)
+  output = lsp_add_stars(x = stars::st_as_stars(x), window = window)
   output_names = names(output)
   output = terra::rast(output)
   names(output) = output_names